pyproject.toml
[project]
name = "credit-underwriting-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.94.0",
    "python-dotenv>=1.2.2",
]
